Modularization of task file
