+++ nextjs/src/components/Task/Models.tsx
@@ -1,0 +1,28 @@
+export interface Question {
+    id: string;
+    type: string;
+    title: string;
+    mission: number;
+    difficulty: number;
+    options: string[] | string[][];
+}
+export interface Answer {
+    id: string;
+    title: string;
+    option: string | number[];
+}
+
+export interface RevisionData {
+    questions: Question[];
+    hits: number[];
+}
+
+export interface TableData {
+    question: Question;
+}
+
+export interface TaskData {
+    username: string, 
+    quests: Question[], 
+    finishFunc: (value: number) => void 
+}
