+++ nextjs/src/components/Task/RadioTable.tsx
@@ -1,0 +1,15 @@
+import { TableData } from './Models';
+
+export default function RadioTable({ question }: TableData) {
+    return(<>{
+    question.options.map(option => (
+        <label key = {option}>
+            <input 
+                type="radio" 
+                name = {question.id}
+                value = {option}/>
+            <span></span>
+            <p> {option} </p>
+        </label>
+    ))}</>)
+}
