+++ nextjs/tsconfig.json
@@ -16,7 +16,11 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "jsx": "preserve"
+    "jsx": "preserve",
+    "baseUrl": "./src",
+    "paths": {
+      "@/public/*": ["../public/*"]
+    }
   },
   "include": [
     "next-env.d.ts",
